pyproject.toml
[project]
name = "JSON-Resume"
version = "0.1.1"
description = "Default template for PDM package"
authors = [
    {name = "Patrick Allan", email = "patrick29501@gmx.com"},
]
dependencies = ["jsons>=1.6.3", "pyyaml>=6.0.2", "toml>=0.10.2"]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm]
distribution = false

[dependency-groups]
dev = [
    "PDFKit>=1.0.0",
    "ReportLab>=4.4.0",
    "pypdf[full]>=5.5.0",
    "pdfminer-six[image]>=20250506",
]
